% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crypton1.R
\name{crypton}
\alias{crypton}
\title{The Crypton Game}
\usage{
crypton(...)
}
\arguments{
\item{...}{\code{crypton} function is called by different arguments, which vary depending
on a problem that Bit is trying to solve. See \code{Details} in order to learn more about the list of possible arguments.}
}
\description{
The \code{crypton} function ...
}
\details{
Every time when some additional hints are needed one should add
\code{hint=TRUE} argument to the \code{crypton} function.
}
\examples{
\dontrun{
crypton()
crypton(hint=TRUE)
}
}
\author{
Tymoteusz Wolodzko (crypton game),
Przemyslaw Biecek (documentation / man page)
}

